refactor(daemon): extract the permission/elicitation coordinator into src/permissions/
Move the ACP `session/request_permission` + `elicitation/create` cluster off `Daemon`
into `PermissionCoordinator` (src/permissions/coordinator.ts), verbatim, together with
the four maps it exclusively owns. Per the delegate-cut policy no same-name shells stay
behind: every call site now holds the coordinator directly via `this.permissions.*`.
